refactor(web): move screens to  page.tsx and structural changes
