--- packages/web/screens/signup.tsx
@@ -1,239 +1,0 @@
-"use client";
-
-import { zodResolver } from "@hookform/resolvers/zod";
-import Link from "next/link";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
-import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { Separator } from "@/components/ui/separator";
-import { Spinner } from "@/components/ui/spinner";
-import Image from "next/image";
-import axios from "axios";
-import { useState } from "react";
-import { useRouter } from "next/navigation";
-import { signIn } from "next-auth/react";
-import { toast } from "sonner";
-
-const formSchema = z.object({
-  name: z.string().min(2, "Name must be at least 2 characters long"),
-  email: z.string().email("Please enter a valid email address"),
-  password: z.string().min(8, "Password must be at least 8 characters long"),
-});
-
-const SignUp = () => {
-  const router = useRouter();
-  const [message, setMessage] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const form = useForm<z.infer<typeof formSchema>>({
-    defaultValues: {
-      name: "",
-      email: "",
-      password: "",
-    },
-    resolver: zodResolver(formSchema),
-  });
-
-  const api = axios.create({
-    baseURL: process.env.NEXT_PUBLIC_BACKEND,
-  });
-
-  const onSubmit = async (data: z.infer<typeof formSchema>) => {
-    setIsLoading(true);
-    try {
-      const res = await api.post("/auth/signup", data);
-      if (res.status === 201) {
-        toast.success("Verification email sent", {
-          description: "Please check your inbox and verify your account.",
-        });
-
-        setTimeout(() => {
-          router.push("/login?verify=sent");
-        }, 2000);
-      }
-    } catch (error: any) {
-      if (error.response) {
-        setMessage(error.response.data.msg);
-      } else {
-        setMessage("Something went wrong. Please try again.");
-      }
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <div className="flex min-h-screen items-center justify-center bg-muted/40 p-4">
-      <Card className="w-full max-w-md">
-        <CardHeader className="space-y-1 text-center">
-          <div className="mx-auto mb-4 flex items-center justify-center ">
-            <Image src="/logo.svg" alt="XCHG logo" width={200} height={50} />
-          </div>
-          <CardTitle className="text-2xl font-bold">
-            Create an account
-          </CardTitle>
-          <CardDescription>
-            Enter your details below to create your account
-          </CardDescription>
-        </CardHeader>
-
-        <CardContent className="space-y-4">
-          <Form {...form}>
-            <form className="space-y-4" onSubmit={form.handleSubmit(onSubmit)}>
-              <FormField
-                control={form.control}
-                name="name"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Full Name</FormLabel>
-                    <FormControl>
-                      <Input
-                        placeholder="John Doe"
-                        type="text"
-                        autoComplete="name"
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="email"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Email</FormLabel>
-                    <FormControl>
-                      <Input
-                        placeholder="name@example.com"
-                        type="email"
-                        autoComplete="email"
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="password"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Password</FormLabel>
-                    <FormControl>
-                      <Input
-                        placeholder="Create a password"
-                        type="password"
-                        autoComplete="new-password"
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <Button className="w-full" type="submit" disabled={isLoading}>
-                {isLoading && <Spinner className="mr-2 h-4 w-4" />}
-                Create account
-              </Button>
-            </form>
-          </Form>
-
-          <div className="relative">
-            <div className="py-2 absolute inset-0 flex items-center">
-              <Separator />
-            </div>
-            <div className="relative flex justify-center text-xs uppercase">
-              <span className="bg-card px-2 text-muted-foreground">Or</span>
-            </div>
-          </div>
-
-          <Button
-            variant="outline"
-            className="w-full"
-            type="button"
-            onClick={() => signIn("github", { callbackUrl: "/" })}>
-            <GitHubLogo />
-            Continue with Github
-          </Button>
-        </CardContent>
-
-        <CardFooter className="flex flex-col space-y-4">
-          <div className="text-center text-base text-muted-foreground">
-            {message && <p className={"text-red-500"}>{message}</p>}
-          </div>
-          <div className="text-center text-sm text-muted-foreground">
-            Already have an account?{" "}
-            <Link
-              href="/login"
-              className="font-medium text-primary underline-offset-4 hover:underline">
-              Sign in
-            </Link>
-          </div>
-
-          <div className="text-center text-xs text-muted-foreground">
-            By clicking continue, you agree to our{" "}
-            <Link
-              href="/terms"
-              className="underline underline-offset-4 hover:text-primary">
-              Terms of Service
-            </Link>{" "}
-            and{" "}
-            <Link
-              href="/privacy"
-              className="underline underline-offset-4 hover:text-primary">
-              Privacy Policy
-            </Link>
-            .
-          </div>
-        </CardFooter>
-      </Card>
-    </div>
-  );
-};
-
-const GitHubLogo = () => (
-  <svg
-    className="mr-2 h-4 w-4"
-    viewBox="0 0 16 16"
-    fill="currentColor"
-    xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M8 0C3.58 0 0 3.58 0 8a8 8 0 005.47 7.59c.4.07.55-.17.55-.38
-      0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13
-      -.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87
-      2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95
-      0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21
-      2.2.82a7.66 7.66 0 012.01-.27c.68 0 1.36.09 2.01.27
-      1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08
-      2.12.51.56.82 1.27.82 2.15 0 3.07-1.87
-      3.75-3.65 3.95.29.25.54.73.54 1.48
-      0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.01
-      8.01 0 0016 8c0-4.42-3.58-8-8-8z"
-    />
-  </svg>
-);
-
-export default SignUp;
